% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart}
\alias{bart}
\title{Calculates index scores for BART game specifically}
\usage{
bart(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{mean_hits}{Mean of hits for balloons not exploded.}
\item{mean_hits_raw}{Mean of hits for all balloons.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
The adjusted and unadjusted BART scores are both returned.
}
