% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps.R
\name{bps}
\alias{bps}
\title{Calculates index scores for Choice Reaction Time game}
\usage{
bps(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{bps_score}{BPS score.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Now we majorly consider two indices: mean reaction time and count of correct
responses.
}
