% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{calc_switch_cost}
\alias{calc_switch_cost}
\title{Switch cost}
\usage{
calc_switch_cost(
  data,
  name_block = "Block",
  name_task = "Task",
  name_switch = "TaskType",
  name_rt = "RT",
  name_acc = "ACC",
  values_pure = c("", "Pure"),
  values_mixed = c("Repeat", "Switch"),
  dur_pure = 0.5,
  dur_mixed = 1
)
}
\arguments{
\item{data}{A \code{data.frame} from which the indices are calculated.}

\item{name_block}{The name of the variable in \code{data} storing block number.}

\item{name_task}{The name of variable in \code{data} storing task info.}

\item{name_switch}{The name of the variable in \code{data} storing switch info.}

\item{name_rt}{The name of the variable in \code{data} storing reaction time data.}

\item{name_acc}{The name of the variable in \code{data} storing accuracy data.}

\item{values_pure}{The values of the variable \code{name_switch} in pure blocks.}

\item{values_mixed}{The values of the variable \code{name_switch} in mixed
blocks, in which the first is about 'repeat', and the second about
'switch'.}

\item{dur_pure}{The duration in minutes for pure blocks.}

\item{dur_mixed}{The duration in minutes for mixed blocks.}
}
\description{
Utility function to calculate indices related to switch cost.
}
