% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexswitch.R
\name{complexswitch}
\alias{complexswitch}
\title{Calculates index scores for complex switch games}
\usage{
complexswitch(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:
\item{mrt_con}{Mean reaction time for congruent trials.}
\item{mrt_inc}{Mean reaction time for incogruent trials.}
\item{cong_eff_rt}{Congruency effect of reaction time (RT), i.e., RT
incongruency - RT congruency.}
\item{pc_con}{Percent of correct for congruent trials.}
\item{pc_inc}{Percent of correct for incogruent trials.}
\item{cong_eff_pc}{Congruency effect of percent of correct (PC), i.e., PC
congruency - PC incongruency.}
\item{rc_mixed}{Count of correct responses per minute for mixed blocks.}
\item{rc_pure}{Count of correct responses per minute for pure blocks.}
\item{switch_cost_rc_gen}{General switch cost (based on count of correct
responses).}
\item{mrt_pure}{Mean reaction time for non-mixed blocks.}
\item{mrt_repeat}{Mean reaction time for repeat trials.}
\item{mrt_switch}{Mean reaction time for switch trials.}
\item{switch_cost_gen_rt}{General switch cost (based on mean reation
times).}
\item{switch_cost_spe_rt}{Specific switch cost (based on mean reation
times).}
\item{nc}{Count of correct responses.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
This is just to find out the count of correct responses.
}
