% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congeff.R
\name{congeff}
\alias{congeff}
\title{Calculates index scores for games related to congruency effect}
\usage{
congeff(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{mrt_inc}{Mean reaction time for incogruent trials.}
\item{mrt_con}{Mean reaction time for congruent trials.}
\item{cong_eff_rt}{Congruency effect of reaction time (RT), i.e., RT
incongruency - RT congruency.}
\item{pc_inc}{Percent of correct for incogruent trials.}
\item{pc_con}{Percent of correct for congruent trials.}
\item{cong_eff_pc}{Congruency effect of percent of correct (PC), i.e., PC
congruency - PC incongruency.}
\item{nc}{Count of correct responses.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
Count of correct responses and congruency effect of correct response and
reaction time are all included.
}
