% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countcorrect.R
\name{countcorrect}
\alias{countcorrect}
\title{Calculates index scores for games only considering correct response numbers.}
\usage{
countcorrect(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{nc}{Count of correct responses.}
\item{pc}{Percent of correct responses.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
This is just to find out the count of correct responses.
}
