% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt}
\alias{cpt}
\title{Calculates index scores for Continuous Performance Test game.}
\usage{
cpt(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{dprime}{Sensitivity (d').}
  \item{c}{Bias index.}
  \item{hits}{Number of hits.}
  \item{commissions}{Number of errors caused by action.}
  \item{omissions}{Number of errors caused by inaction.}
  \item{count_error}{Count of incorrect responses.}
  \item{mrt}{Mean reaction time of hits.}
  \item{rtsd}{Standard deviation of reaction times of hits.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Many indices are returned: d', c (i.e., bias), hits, commissions, ommissions,
mean reaction time (mrt), standard deviation of reaction times (rtsd).
}
