% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt.R
\name{crt}
\alias{crt}
\title{Calculates index scores for Choice Reaction Time game}
\usage{
crt(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{mrt}{Mean reaction time}
  \item{count_correct}{Count of correct responses}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Now we majorly consider two indices: mean reaction time and count of correct
responses.
}
