% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drm.R
\name{drm}
\alias{drm}
\title{Calculates index scores for DRM paradigm game}
\usage{
drm(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{fm_ratio}{False memory ratio.}
  \item{fm_dprime}{False memory d'.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Two major indices are included: fm_ratio (false memory ratio, i.e.,
p(old|lure) - p(old|foil)) and fm_dprime (false memory d', i.e.,
z(p(old|lure)) - z(p(old|foil)))
}
