% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualtask.R
\name{dualtask}
\alias{dualtask}
\title{Calculates index scores for dual task paradigm}
\usage{
dualtask(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{mrt_all}{Mean reaction time for all correct trials.}
\item{mrt_left}{Mean reaction time for correct trials of left side.}
\item{mrt_right}{Mean reaction time for correct trials of right side.}
\item{nc_all}{Number of correct responses of both sides.}
\item{nc_left}{Number of correct responses of left side.}
\item{nc_right}{Number of correct responses of right side.}
\item{dprime_all}{Sensitivity index (d') of both sides.}
\item{dprime_left}{Sensitivity index (d') of left side.}
\item{dprime_right}{Sensitivity index (d') of right side.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
In the simplified (only dual task block included) dual task paradigm, users
are required to complete two tasks spontaneously, one at the left side, and
the other at the right side. Mean reaction time (\code{mrt}), count of
correct responses (\code{nc}) and sensitivity index (\code{dprime}) for left side,
right side and both side are calculated.
}
