% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlo.R
\name{jlo}
\alias{jlo}
\title{Calculates index scores for JLO game.}
\usage{
jlo(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{count_correct}{Count of correct responses.}
\item{sum_error}{Sum of the angle deviations.}
\item{sum_logerr}{Sum of the log of angle deviations.}
\item{sum_sqrterr}{Sum of the square root of angle deviations.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
Count of correct responses and other angle deviations indices.
}
