% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london.R
\name{london}
\alias{london}
\title{Calculates index scores for London Tower game.}
\usage{
london(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{tscore}{Total score defined by the game itself.}
  \item{mean_level}{Mean level reached.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Two indices are returned: total scores (defined in the game itself) and mean
level.
}
