% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisense.R
\name{multisense}
\alias{multisense}
\title{Calculates index scores for Multisense games.}
\usage{
multisense(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{mrt_image}{Mean reaction time of Image stimuli.}
  \item{mrt_sound}{Mean reaction time of Sound stimuli.}
  \item{mrt_mixed}{Mean reaction time of Mixed stimuli.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Three mean reaction times as to Image, Sound and Mixed stimuli are returned.
}
