% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nback.R
\name{nback}
\alias{nback}
\title{Calculates index scores for N Back games.}
\usage{
nback(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{pc}{Percentage of correct responses.}
\item{mrt}{Mean reaction time.}
\item{dprime}{Sensitivity index.}
\item{c}{Bias.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
Several values including percentage of correct responses (pc), mean reaction
time (mrt), sensitivity index (i.e, d') and bias (c).
}
