% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nle.R
\name{nle}
\alias{nle}
\title{Calculates index scores for Number Line Estimation game.}
\usage{
nle(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{mean_err}{Mean absolute error.}
  \item{mean_logerr}{Mean log absolute error.}
  \item{mean_sqrterr}{Mean square root of absolute error.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Now the mean absolute error (mean_err) is calculated. Future work will
be to do model fitting.
}
