% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsymncmp.R
\name{nsymncmp}
\alias{nsymncmp}
\title{Calculates index scores for Nonsymbolic Number Comparison game.}
\usage{
nsymncmp(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{pc}{Percentage of correct responses.}
\item{mrt}{Mean reaction time.}
\item{w}{Weber fraction.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
Several values including percentage of correct responses (pc), mean reaction
time (mrt), weber fraction (w).
}
