% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schulte.R
\name{schulte}
\alias{schulte}
\title{Calculates index scores for Schulte Grid games}
\usage{
schulte(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{net_cor}{Net correct count.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
The net count of correct responses (net_cor) is returned, which is just the
difference of the count of correct responses and error responses.
}
