% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{span}
\alias{span}
\title{Calculates index scores for Span related games}
\usage{
span(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{max_span}{Maximal span.}
\item{mean_span}{Mean span.}
\item{count_correct}{Count of correct responses.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
Three indices are returned: maximal span, mean span and count of correct
responses.
}
