% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srt.R
\name{srt}
\alias{srt}
\title{Calculates index scores for Simple Reaction Time game}
\usage{
srt(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{mrt}{Mean reaction time}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
Mean reaction time is returned.
}
