% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchcost.R
\name{switchcost}
\alias{switchcost}
\title{Calculates index scores for games related to conflict effect}
\usage{
switchcost(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{count_correct}{Count of correct responses.}
\item{count_pure}{Count of correct responses per minute for pure blocks.}
\item{count_mixed}{Count of correct responses per minute for mixed blocks.}
\item{switch_cost_gen_count}{General switch cost (based on count of correct responses).}
\item{mrt_pure}{Mean reaction time for non-mixed blocks.}
\item{mrt_repeat}{Mean reaction time for repeat trials.}
\item{mrt_switch}{Mean reaction time for switch trials.}
\item{switch_cost_gen_rt}{General switch cost (based on mean reation times).}
\item{switch_cost_spe_rt}{Specific switch cost (based on mean reation times).}
\item{is_normal}{Checking result whether the data is normal.} }
}
\description{
Count of correct responses and congruency effect of correct response and
reaction time are all included.
}
