% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symncmp.R
\name{symncmp}
\alias{symncmp}
\title{Calculates index scores for Symbolic Number Comparison game.}
\usage{
symncmp(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{pc}{Percentage of correct responses.}
  \item{mrt}{Mean reaction time.}
  \item{dist_eff}{Distance effect.}
  \item{dist_eff_adj}{Adjusted distance effect.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Several values including percentage of correct responses (pc), mean reaction
time (mrt), distance effect (dist_effect) and adjusted distance effect
(dist_effect_adj).
}
