% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{calc_switch_cost}
\alias{calc_switch_cost}
\title{Switch cost}
\usage{
calc_switch_cost(
  data,
  config_block,
  name_block = "Block",
  name_task = "Task",
  name_switch = "Type",
  name_rt = "RT",
  name_acc = "ACC",
  values_mixed = c("Repeat", "Switch")
)
}
\arguments{
\item{data}{Required. A \code{data.frame} from which the indices are calculated.}

\item{config_block}{Required. A \code{data.frame} contains the configurations of
each block. At least contains these variables:
\itemize{
\item Same name as \code{name_block} argument: The block identifier.
\item \code{has_no_response}: Logical value indicating whether this block has no
valid responses or not.
\item \code{type_block}: The type of each block: "pure" or "mixed".
\item \code{dur}: The duration in minutes for each block.
}}

\item{name_block}{The name of the variable in \code{data} storing block number.}

\item{name_task}{The name of variable in \code{data} storing task info.}

\item{name_switch}{The name of the variable in \code{data} storing switch info.}

\item{name_rt}{The name of the variable in \code{data} storing reaction time data.}

\item{name_acc}{The name of the variable in \code{data} storing accuracy data.}

\item{values_mixed}{The values of the variable \code{name_switch} in mixed
blocks, in which the first is about 'repeat', and the second about
'switch'.}
}
\description{
Utility function to calculate indices related to switch cost.
}
