% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict.R
\name{conflict}
\alias{conflict}
\title{Calculates index scores for games related to conflict effect}
\usage{
conflict(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values: \describe{
  \item{count_correct}{Count of correct responses.}
  \item{cong_eff_rt}{Congruency effect of reaction time (RT), i.e., RT
  incongruency - RT congruency.} \item{cong_eff_pc}{Congruency effect of
  percent of correct (PC), i.e., PC congruency - PC incongruency.}
  \item{is_normal}{Checking result whether the data is normal.} }
}
\description{
Count of correct responses and congruency effect of correct response and
reaction time are all included.
}
