% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlo.R
\name{jlo}
\alias{jlo}
\title{Calculates index scores for JLO game.}
\usage{
jlo(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{count_correct}{Count of correct responses.}
  \item{sum_error}{Sum of the angle deviations.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
This is just to find out the count of correct responses.
}
