% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london.R
\name{london}
\alias{london}
\title{Calculates index scores for London Tower game.}
\usage{
london(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\item{tscore}{Total score defined by the game itself.}
\item{ratio_score}{Mean of the ratio between best steps and actual steps.}
\item{mean_level}{Mean level reached.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
Two indices are returned: total scores (defined in the game itself), ratio
score (defined as the mean of the ratios between best steps and actual steps)
and mean level.
}
