% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxpred.R
\name{wxpred}
\alias{wxpred}
\title{Calculates index for Weather Prediction Paradigm game}
\usage{
wxpred(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \code{data.frame} contains following values:
\describe{
  \item{pc_b1}{Percent of correct responses.}
  \item{is_normal}{Checking result whether the data is normal.}
}
}
\description{
Percent of correct for each block is returned.
}
